Implemented the segment processing, need sentinels
Implemented a function to process the segments into an index and length
for the longest common substring. Sentinels still need to be distinct
from the rest of the bytestring, which means I need to convert the input
into a larger format (2x the space?) to be reasonable. Additionally, the
code could use some cleanup, and the input needs to be collected from
the command line files instead of being hardcoded strings. Similarly, we
need another scan to find the position of the string within each
document, instead of in only one.