Use sentinels that are part of same alphabet (u8)
Switched from `suffix` to `suffix_array` in order to generate suffix
arrays for a bytestring. `suffix` requires a string, and rust's encoding
of strings in UTF-8 implies that it will not work well.

Considered different approaches for unique sentinels, and realized that
the actual value in the array for the sentinel is not necessary, as long
as we keep track of their placement and can check it in O(1).
Implemented the approach I thought of, and it seems to work.

It still needs to be seen if its better to keep the array positions -> doc
index, or if it's better to keep a set (or maybe array if there are few enough)
of sentinel positions, as we are much more frequently checking sentinel
positions than document id. That said, the first step is making it
actually work on file input instead of hardcoded input.